feat: BSZ-FRUN-01-runtime-session-surfaces-007 - Preserve Factory Session HTTP lifecycle surface under file size limit
